Initial import of dtb-arm-vexpress.
The devicetree is a data structure for describing hardware. Rather than
hard coding every detail of a device into an operating system, many aspects
of the hardware can be described in a data structure that is passed to the
operating system at boot time.

This package includes Device Tree Blob (.dtb) files generated from
the Linux kernel for ARM Ltd. Versatile Express boards.