fake-pc.mk: introduce package settable SKIP_FAKE_PC_FILES variable
fake-pc framework create fake pc files in buildlink tree if builtin one
is missing.

It is usuall useful but sometimes unwanted behavior.  For example,
It is supported to build a package without the pc file, and wanted to avoid
such fake pc is recorded as "Requires" in installed pc files from the package.

SKIP_FAKE_PC_FILES is list of pc file names for skipping to create fake pc
files.