chat: Remove vicq.
Upstream is gone, and there have been numerous changes to the ICQ protocol
that rendered third-party clients non-functional since this was last updated.

There is an archive on GitHub which states:

"This is just an archive of very old project. It's not supported,
not functional and has no other value except sentimental"