py-sniffio: added version 1.1.0
You're writing a library. You've decided to be ambitious, and support multiple
async I/O packages, like Trio, and asyncio, and... You've written a bunch of
clever code to handle all the differences. But... how do you know which piece of
clever code to run?

This is a tiny package whose only purpose is to let you detect which async
library your code is running under.